style: rename behaviour
