+++ test/TokenRecover.behavior.js
