+++ test/examples/TRDeployerOwner.test.js
@@ -1,6 +1,6 @@
 const { expect } = require('chai');
 
-const { shouldBehaveLikeTokenRecover } = require('../TokenRecover.behaviour');
+const { shouldBehaveLikeTokenRecover } = require('../TokenRecover.behavior');
 
 const TokenRecover = artifacts.require('$TRDeployerOwner');
